// types.ts
export interface Database {
  public: {
    Tables: {
      admins: {
        Row: {
          id: string;
          created_at: string;
          user_id: string;
        };
        Insert: {
          id?: string;
          created_at?: string;
          user_id?: string;
        };
        Update: {
          id?: string;
          created_at?: string;
          user_id?: string;
        };
      };
      categories: {
        Row: {
          id: string;
          created_at: string;
          title: string;
          slug: string;
          image: string;
        };
        Insert: {
          id?: string;
          created_at?: string;
          title: string;
          slug: string;
          image?: string;
        };
        Update: {
          id?: string;
          created_at?: string;
          title?: string;
          slug?: string;
          image?: string;
        };
      };
// ...
      product_categories: {
        Row: {
          category_id: Database["public"]["Tables"]["categories"]["Row"];
          product_id: Database["public"]["Tables"]["products"]["Row"];
          created_at: string;
        };
        Insert: {
          category_id?: string;
          product_id?: string;
          created_at?: string;
        };
        Update: {
          category_id?: string;
          product_id?: string;
          created_at?: string;
        };
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      get_total_paid_orders: {
        Args: Record<PropertyKey, never>;
        Returns: number;
      };
      is_admin: {
        Args: Record<PropertyKey, never>;
        Returns: boolean;
      };
    };
    Enums: {
      [_ in never]: never;
    };
  };
}